+++ cassandra-mesos-framework/src/main/java/io/mesosphere/mesos/frameworks/cassandra/framework/Main.java
@@ -18,7 +18,9 @@ package io.mesosphere.mesos.frameworks.cassandra.framework;
 import com.google.common.base.Optional;
 import com.google.common.base.Supplier;
 import io.mesosphere.mesos.frameworks.cassandra.scheduler.*;
-import io.mesosphere.mesos.frameworks.cassandra.scheduler.files.FileResourceController;
+import io.mesosphere.mesos.frameworks.cassandra.scheduler.api.ApiController;
+import io.mesosphere.mesos.frameworks.cassandra.scheduler.api.ApiControllerFactory;
+import io.mesosphere.mesos.frameworks.cassandra.scheduler.api.FileResourceController;
 import io.mesosphere.mesos.frameworks.cassandra.scheduler.util.Env;
 import io.mesosphere.mesos.util.Clock;
 import io.mesosphere.mesos.util.ProtoUtils;
@@ -186,8 +188,7 @@ public final class Main {
         );
 
         final ResourceConfig rc = new ResourceConfig()
-            .register(new FileResourceController(cassandraVersion))
-            .register(new ApiController(cassandraCluster));
+            .registerInstances(ApiControllerFactory.buildInstances(cassandraCluster, cassandraVersion));
             //.packages("io.mesosphere.mesos.frameworks.cassandra");
         final HttpServer httpServer = GrizzlyHttpServerFactory.createHttpServer(httpServerBaseUri, rc);
